LPD-35776 Move ObjectFolderCardHeader and ObjectFoldersSidebar frontend unit test files to ViewObjectDefinitions folder
